requirements.txt
meilisearch-python-sdk
mysql-connector-python
tqdm
